Jenkins jobs - JUnit test split into two steps

+++ etc/jenkins/build.groovy
@@ -139,7 +139,10 @@ spec:
                                 mv jpa/org.eclipse.persistence.jpa.jpql.test/reports/TESTS-TestSuites.xml ${HOME}/test.reports/jpa.jpql
                             fi
                             #Move aggregated test report into another location. It duplicated with modules test reports.
-                            mv target/reports/TESTS-TestSuites.xml ${HOME}/test.reports/                            
+                            if [ -f "target/reports/TESTS-TestSuites.xml" ]; then
+                                mkdir -p ${HOME}/test.reports
+                                mv target/reports/TESTS-TestSuites.xml ${HOME}/test.reports/
+                            fi                            
                         """
                     }
                 }
